Copy paste aimandrange into own subproject
