+++ photonlib-cpp-examples/aimandrange/simgui-window.json
@@ -19,7 +19,12 @@
     "###Joysticks": {
       "Collapsed": "0",
       "Pos": "250,465",
-      "Size": "796,73"
+      "Size": "796,206"
+    },
+    "###Keyboard 0 Settings": {
+      "Collapsed": "0",
+      "Pos": "10,50",
+      "Size": "300,560"
     },
     "###NetworkTables": {
       "Collapsed": "0",
